# shared/openapi/spec.yaml
openapi: 3.0.3
info:
  title: Ubik Enterprise - AI Agent Management Platform
  description: |
    Centralized platform for managing employee AI agent and MCP configurations.
    
    Companies can:
    - Manage employees, teams, and roles
    - Configure AI agents (Claude Code, Cursor, Windsurf)
    - Configure MCP servers
    - Set policies and approval workflows
    - Track usage and costs
  version: 1.0.0
  contact:
    name: API Support
    email: support@ubik-enterprise.com
servers:
  - url: /api/v1
    description: Current server
...
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    Error:
      type: object
      required:
        - error
        - message
      properties:
        error:
          type: string
          example: "validation_error"
        message:
          type: string
          example: "Email is required"
        details:
          type: object
          additionalProperties: true

    # Health check
...
    CreateInvitationRequest:
      type: object
      required:
        - email
        - role_id
      properties:
        email:
          type: string
          format: email
          description: Email address of the person being invited
        role_id:
          type: string
          format: uuid
          description: Role to assign to the new employee
        team_id:
          type: string
          format: uuid
          nullable: true
          description: Optional team to assign the employee to
    Invitation:
      type: object
      required:
        - id
        - org_id
        - inviter_id
        - email
        - role_id
        - token
        - status
        - expires_at
        - created_at
        - updated_at
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        org_id:
          type: string
          format: uuid
        inviter_id:
          type: string
          format: uuid
          description: Employee who created the invitation
        email:
          type: string
          format: email
        role_id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
          nullable: true
        token:
          type: string
          minLength: 64
          maxLength: 64
          description: Secure invitation token (64 character hex string)
        status:
          type: string
          enum: [pending, accepted, expired, cancelled]
          description: Current invitation status
        expires_at:
          type: string
          format: date-time
          description: When the invitation expires (7 days from creation)
        accepted_by:
          type: string
          format: uuid
          nullable: true
          description: Employee ID who accepted the invitation
        accepted_at:
          type: string
          format: date-time
          nullable: true
          description: When the invitation was accepted
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    InvitationListResponse:
      type: object
      required:
        - invitations
        - total
        - limit
        - offset
      properties:
        invitations:
          type: array
          items:
            $ref: '#/components/schemas/Invitation'
        total:
          type: integer
          description: Total number of invitations
        limit:
          type: integer
          description: Number of results per page
        offset:
          type: integer
          description: Offset for pagination
    InvitationDetails:
      type: object
      required:
        - id
        - org_name
        - email
        - role_name
        - status
        - expires_at
      properties:
        id:
          type: string
          format: uuid
        org_name:
          type: string
          description: Name of the organization
        email:
          type: string
          format: email
        role_name:
          type: string
          description: Name of the role being assigned
        team_name:
          type: string
          nullable: true
          description: Name of the team (if assigned)
        status:
          type: string
          enum: [pending, accepted, expired, cancelled]
        expires_at:
          type: string
          format: date-time
          description: When the invitation expires
    AcceptInvitationRequest:
      type: object
      required:
        - full_name
        - password
      properties:
        full_name:
          type: string
          minLength: 1
          maxLength: 255
          description: Full name of the employee
        password:
          type: string
          minLength: 8
          format: password
          description: Password (must be at least 8 characters)
    AcceptInvitationResponse:
      type: object
      required:
        - token
        - expires_at
        - employee
      properties:
        token:
          type: string
          description: JWT authentication token
        expires_at:
          type: string
          format: date-time
          description: Token expiration timestamp
        employee:
          $ref: '#/components/schemas/Employee'

    # Core domain schemas
    Employee:
      type: object
      required:
        - id
        - org_id
        - email
        - full_name
        - role_id
        - status
      properties:
        id:
          type: string
          format: uuid
          readOnly: true
        org_id:
          type: string
          format: uuid
        team_id:
          type: string
          format: uuid
          nullable: true
        team_name:
          type: string
          description: Name of the team (populated in list endpoints)
          readOnly: true
          nullable: true
        role_id:
          type: string
          format: uuid
        email:
          type: string
          format: email
          maxLength: 255
        full_name:
          type: string
          minLength: 1
          maxLength: 255
        status:
          type: string
          enum: [active, suspended, inactive]
          default: active
        preferences:
          type: object
          additionalProperties: true
        has_personal_claude_token:
          type: boolean
          description: Whether employee has configured a personal Claude API token
          readOnly: true
        last_login_at:
          type: string
          format: date-time
          nullable: true
          readOnly: true
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
    CreateEmployeeRequest:
      type: object
      required:
        - email
        - full_name
        - role_id
      properties:
        email:
          type: string
          format: email
          maxLength: 255
        full_name:
          type: string
          minLength: 1
          maxLength: 255
        team_id:
          type: string
          format: uuid
          nullable: true
        role_id:
          type: string
          format: uuid
        preferences:
          type: object
          additionalProperties: true
          default: {}
    CreateEmployeeResponse:
      type: object
      required:
        - employee
        - temporary_password
      properties:
        employee:
          $ref: '#/components/schemas/Employee'
        temporary_password:
          type: string
          description: Temporary password for the new employee (only returned on creation)
          example: "aB3$xY9#mK2@pL7!"
          minLength: 16
          maxLength: 32
...
    ListEmployeesResponse:
      type: object
      required:
        - employees
        - total
        - limit
        - offset
      properties:
        employees:
          type: array
          items:
            $ref: '#/components/schemas/Employee'
        total:
          type: integer
          format: int64
          description: Total number of employees matching the filter
        limit:
          type: integer
          description: Maximum number of results per page
        offset:
          type: integer
          description: Number of results to skip
...
  parameters:
...
    Page:
      name: page
      in: query
      schema:
        type: integer
        minimum: 1
        default: 1
      description: Page number
    PerPage:
      name: per_page
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 20
      description: Items per page
    Limit:
      name: limit
      in: query
      schema:
        type: integer
        minimum: 1
        maximum: 100
        default: 10
      description: Number of items to return
    Offset:
      name: offset
      in: query
      schema:
        type: integer
        minimum: 0
        default: 0
      description: Number of items to skip
    Status:
      name: status
      in: query
      schema:
        type: string
        enum: [active, suspended, inactive]
      description: Filter by status
paths:
...
  /invitations:
    post:
      tags:
        - invitations
      summary: Create invitation
      description: Create an invitation for a new employee (admin only, rate limited to 20/day)
      operationId: createInvitation
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateInvitationRequest'
      responses:
        '201':
          description: Invitation created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Invitation'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - requires admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Rate limit exceeded (20 invitations/day)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
    get:
      tags:
        - invitations
      summary: List invitations
      description: Get paginated list of invitations for current organization (admin only)
      operationId: listInvitations
      parameters:
        - $ref: '#/components/parameters/Limit'
        - $ref: '#/components/parameters/Offset'
      responses:
        '200':
          description: List of invitations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - requires admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /invitations/{token}:
    get:
      tags:
        - invitations
      summary: Get invitation by token
      description: Validate invitation token and get details (public endpoint)
      operationId: getInvitationByToken
      security: []  # No auth required - public endpoint
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            minLength: 64
            maxLength: 64
          description: Invitation token (64 character hex string)
      responses:
        '200':
          description: Invitation details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InvitationDetails'
        '404':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '410':
          description: Invitation expired or cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /invitations/{token}/accept:
    post:
      tags:
        - invitations
      summary: Accept invitation
      description: Accept invitation and create employee account (public endpoint)
      operationId: acceptInvitation
      security: []  # No auth required - public endpoint
      parameters:
        - name: token
          in: path
          required: true
          schema:
            type: string
            minLength: 64
            maxLength: 64
          description: Invitation token
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AcceptInvitationRequest'
      responses:
        '200':
          description: Invitation accepted, employee created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AcceptInvitationResponse'
        '400':
          description: Invalid request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already exists or invitation already accepted
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '410':
          description: Invitation expired or cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /invitations/{id}:
    delete:
      tags:
        - invitations
      summary: Cancel invitation
      description: Cancel a pending invitation (admin only)
      operationId: cancelInvitation
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
            format: uuid
          description: Invitation ID
      responses:
        '204':
          description: Invitation cancelled successfully
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - requires admin role
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Invitation not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Cannot cancel accepted invitation
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # ============================================================================
  # Employees
  # ============================================================================
  /employees:
    get:
      tags:
        - employees
      summary: List employees
      description: Get paginated list of employees in current organization
      operationId: listEmployees
      parameters:
        - $ref: '#/components/parameters/Page'
        - $ref: '#/components/parameters/PerPage'
        - $ref: '#/components/parameters/Status'
      responses:
        '200':
          description: Employee list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListEmployeesResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    post:
      tags:
        - employees
      summary: Create employee
      description: Create new employee (admin/manager only)
      operationId: createEmployee
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateEmployeeRequest'
      responses:
        '201':
          description: Employee created with temporary password
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CreateEmployeeResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '403':
          description: Forbidden - insufficient permissions
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '422':
          description: Validation error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...